+++ src/components/DateInput.jsx
@@ -2,9 +2,8 @@ import React from 'react';
 import PropTypes from 'prop-types';
 import { forbidExtraProps } from 'airbnb-prop-types';
 import cx from 'classnames';
-import throttle from 'lodash.throttle';
-
-import isTouchDevice from '../utils/isTouchDevice';
+import throttle from 'lodash/throttle';
+import isTouchDevice from 'is-touch-device';
 
 const propTypes = forbidExtraProps({
   id: PropTypes.string.isRequired,
@@ -38,7 +37,7 @@ const defaultProps = {
   focused: false,
   disabled: false,
   required: false,
-  readOnly: false,
+  readOnly: null,
   showCaret: false,
 
   onChange() {},
@@ -173,7 +172,7 @@ export default class DateInput extends React.Component {
           placeholder={placeholder}
           autoComplete="off"
           disabled={disabled}
-          readOnly={readOnly || isTouch}
+          readOnly={typeof readOnly === 'boolean' ? readOnly : isTouch}
           required={required}
           aria-describedby={screenReaderMessage && screenReaderMessageId}
         />